refactor: 공통 엔티티 생성
