+++ com.devsquad10.hub/src/main/java/com/devsquad10/hub/domain/model/Base.java
@@ -1,0 +1,55 @@
+package com.devsquad10.hub.domain.model;
+
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+import org.hibernate.annotations.SQLRestriction;
+import org.springframework.data.annotation.CreatedBy;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedBy;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import jakarta.persistence.Temporal;
+import jakarta.persistence.TemporalType;
+import lombok.Getter;
+
+@Getter
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+@SQLRestriction("deleted_at IS NULL")
+public class Base {
+
+	@CreatedDate
+	@Column(name = "created_at", nullable = false, updatable = false)
+	@Temporal(TemporalType.TIMESTAMP)
+	private LocalDateTime createdAt;
+
+	@CreatedBy
+	@Column(name = "created_by", nullable = false, updatable = false)
+	private UUID createdBy;
+
+	@LastModifiedDate
+	@Column(name = "updated_at")
+	@Temporal(TemporalType.TIMESTAMP)
+	private LocalDateTime updatedAt;
+
+	@LastModifiedBy
+	@Column(name = "updated_by")
+	private UUID updatedBy;
+
+	@Column(name = "deleted_at")
+	@Temporal(TemporalType.TIMESTAMP)
+	private LocalDateTime deletedAt;
+
+	@Column(name = "deleted_by")
+	private UUID deletedBy;
+
+	public void delete(UUID deletedBy) {
+		this.deletedAt = LocalDateTime.now();
+		this.deletedBy = deletedBy;
+	}
+}
