+++ com.devsquad10.hub/src/main/java/com/devsquad10/hub/domain/model/Hub.java
@@ -1,22 +1,13 @@
 package com.devsquad10.hub.domain.model;
 
-import java.time.LocalDateTime;
 import java.util.UUID;
-
-import org.hibernate.annotations.SQLRestriction;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.PrePersist;
-import jakarta.persistence.PreUpdate;
 import jakarta.persistence.Table;
-import jakarta.persistence.Temporal;
-import jakarta.persistence.TemporalType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -28,10 +19,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Table(name = "p_hub")
-@SQLRestriction("deleted_at IS NULL")
-// TODO: Audit Fields 분리
-@EntityListeners(AuditingEntityListener.class)
-public class Hub {
+public class Hub extends Base {
 	@Id
 	@GeneratedValue(strategy = GenerationType.UUID)
 	private UUID id;
@@ -48,53 +36,10 @@ public class Hub {
 	@Column(nullable = false)
 	private Double longitude;
 
-	// TODO: Audit Fields 분리
-	@Column(name = "created_at", nullable = false, updatable = false)
-	@Temporal(TemporalType.TIMESTAMP)
-	private LocalDateTime createdAt;
-
-	@Column(name = "created_by", nullable = false, updatable = false)
-	private UUID createdBy;
-
-	@Column(name = "updated_at")
-	@Temporal(TemporalType.TIMESTAMP)
-	private LocalDateTime updatedAt;
-
-	@Column(name = "updated_by")
-	private UUID updatedBy;
-
-	@Column(name = "deleted_at")
-	@Temporal(TemporalType.TIMESTAMP)
-	private LocalDateTime deletedAt;
-
-	@Column(name = "deleted_by")
-	private UUID deletedBy;
-
-	@PrePersist
-	protected void onCreate() {
-		this.createdAt = LocalDateTime.now();
-		this.updatedAt = LocalDateTime.now();
-
-		// TODO: created_by
-		this.createdBy = UUID.randomUUID();
-	}
-
-	@PreUpdate
-	protected void onUpdate() {
-		this.updatedAt = LocalDateTime.now();
-
-		// TODO: updated_by
-	}
-
 	public void update(String name, String address, Double latitude, Double longitude) {
 		this.name = name;
 		this.address = address;
 		this.latitude = latitude;
 		this.longitude = longitude;
-	}
-
-	public void delete(UUID deletedBy) {
-		this.deletedAt = LocalDateTime.now();
-		this.deletedBy = deletedBy;
 	}
 }
