+++ com.devsquad10.hub/src/main/java/com/devsquad10/hub/domain/model/HubRoute.java
@@ -1,17 +1,11 @@
 package com.devsquad10.hub.domain.model;
 
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
-import org.hibernate.annotations.SQLRestriction;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
 import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -19,11 +13,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
-import jakarta.persistence.PrePersist;
-import jakarta.persistence.PreUpdate;
 import jakarta.persistence.Table;
-import jakarta.persistence.Temporal;
-import jakarta.persistence.TemporalType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -38,10 +28,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Table(name = "p_hub_route")
-@SQLRestriction("deleted_at IS NULL")
-// TODO: Audit Fields 분리
-@EntityListeners(AuditingEntityListener.class)
-public class HubRoute {
+public class HubRoute extends Base {
 	@Id
 	@GeneratedValue(strategy = GenerationType.UUID)
 	private UUID id;
@@ -62,52 +49,9 @@ public class HubRoute {
 	@Builder.Default
 	private List<HubRouteWaypoint> waypoints = new ArrayList<>();
 
-	// TODO: Audit Fields 분리
-	@Column(name = "created_at", nullable = false, updatable = false)
-	@Temporal(TemporalType.TIMESTAMP)
-	private LocalDateTime createdAt;
-
-	@Column(name = "created_by", nullable = false, updatable = false)
-	private UUID createdBy;
-
-	@Column(name = "updated_at")
-	@Temporal(TemporalType.TIMESTAMP)
-	private LocalDateTime updatedAt;
-
-	@Column(name = "updated_by")
-	private UUID updatedBy;
-
-	@Column(name = "deleted_at")
-	@Temporal(TemporalType.TIMESTAMP)
-	private LocalDateTime deletedAt;
-
-	@Column(name = "deleted_by")
-	private UUID deletedBy;
-
-	@PrePersist
-	protected void onCreate() {
-		this.createdAt = LocalDateTime.now();
-		this.updatedAt = LocalDateTime.now();
-
-		// TODO: created_by
-		this.createdBy = UUID.randomUUID();
-	}
-
-	@PreUpdate
-	protected void onUpdate() {
-		this.updatedAt = LocalDateTime.now();
-
-		// TODO: updated_by
-	}
-
 	public void update(Double distance, Integer duration) {
 		this.distance = distance;
 		this.duration = duration;
-	}
-
-	public void delete(UUID deletedBy) {
-		this.deletedAt = LocalDateTime.now();
-		this.deletedBy = deletedBy;
 	}
 }
