+++ com.devsquad10.hub/src/main/java/com/devsquad10/hub/domain/model/HubRouteWaypoint.java
@@ -1,9 +1,7 @@
 package com.devsquad10.hub.domain.model;
 
-import java.time.LocalDateTime;
 import java.util.UUID;
 
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
@@ -11,11 +9,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.PrePersist;
-import jakarta.persistence.PreUpdate;
 import jakarta.persistence.Table;
-import jakarta.persistence.Temporal;
-import jakarta.persistence.TemporalType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -30,7 +24,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Table(name = "p_hub_route_waypoint")
-public class HubRouteWaypoint {
+public class HubRouteWaypoint extends Base {
 	@Id
 	@GeneratedValue(strategy = GenerationType.UUID)
 	private UUID id;
@@ -54,42 +48,4 @@ public class HubRouteWaypoint {
 	private Double distance;
 
 	private Integer duration;
-
-	// TODO: Audit Fields 분리
-	@Column(name = "created_at", nullable = false, updatable = false)
-	@Temporal(TemporalType.TIMESTAMP)
-	private LocalDateTime createdAt;
-
-	@Column(name = "created_by", nullable = false, updatable = false)
-	private UUID createdBy;
-
-	@Column(name = "updated_at")
-	@Temporal(TemporalType.TIMESTAMP)
-	private LocalDateTime updatedAt;
-
-	@Column(name = "updated_by")
-	private UUID updatedBy;
-
-	@Column(name = "deleted_at")
-	@Temporal(TemporalType.TIMESTAMP)
-	private LocalDateTime deletedAt;
-
-	@Column(name = "deleted_by")
-	private UUID deletedBy;
-
-	@PrePersist
-	protected void onCreate() {
-		this.createdAt = LocalDateTime.now();
-		this.updatedAt = LocalDateTime.now();
-
-		// TODO: created_by
-		this.createdBy = UUID.randomUUID();
-	}
-
-	@PreUpdate
-	protected void onUpdate() {
-		this.updatedAt = LocalDateTime.now();
-
-		// TODO: updated_by
-	}
 }
